requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-etc==1.3.0
django-hitcount==1.3.3
django-js-asset==1.2.2
django-resized==0.3.11
django-taggit==1.3.0
easy-thumbnails==2.7.1
idna==2.10
Pillow==8.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
